BUG: fix snippets for phi, epsilon, theta
I added a `2` to the name of the snippet that converts from `varNAME` to `NAME`. 

Having it named `NAME` masked the original snippet with the exact same name and made the snippets not work at the original entry point (i.e. `e[TAB]` wasn't inserting `\epsilon` because the snippet named `epsilon` with that prefix was masked by the snippet with the same name and prefix `varepsilon`).